% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ames_housing.R
\docType{data}
\name{ames_housing}
\alias{ames_housing}
\alias{mlr_tasks_ames_housing}
\title{House Sales in Ames, Iowa}
\description{
Regression task to predict house sale prices for Ames, Iowa.

Contains 80 features and 2930 observations.
Target column is \code{"Sale_Price"}.
}
\examples{
data("ames_housing", package = "mlr3data")
str(ames_housing)
}
\keyword{data}
