% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_sharing.R
\docType{data}
\name{bike_sharing}
\alias{bike_sharing}
\alias{mlr_tasks_bike_sharing}
\title{Bike Sharing Demand}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/bike+sharing+dataset}
}
\description{
Regression data to predict the total count of bikes rented. Contains
13 features and 17379 observations. Target column is \code{"count"}.
}
\section{Pre-processing}{

\itemize{
\item All columns have been renamed.
\item \code{instant}, \code{"registered"} and \code{"casual"} column have been removed.
\item \code{"season"} and \code{"weather"} have been converted to \code{factor()}.
\item \code{"holiday"} and \code{"working_day"} have been converted to \code{logical()}.
}
}

\examples{
data("bike_sharing", package = "mlr3data")
str(bike_sharing)
}
\keyword{data}
