% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_usage.R
\docType{data}
\name{energy_usage}
\alias{energy_usage}
\title{Power Consumption of Kitchen Appliances in Ames, Iowa}
\description{
Data for power consumption of kitchen appliances in Ames, Iowa.
Extends the \code{ames_housing} data set.

Contains 720 features and 2930 observations.
}
\examples{
data("energy_usage", package = "mlr3data")
str(energy_usage)
}
\keyword{data}
