% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilpd.R
\docType{data}
\name{ilpd}
\alias{ilpd}
\alias{mlr_tasks_ilpd}
\title{Indian Liver Patient Dataset}
\source{
\url{https://www.openml.org/d/1480}
}
\description{
Classification data to predict whether or not a person is a liver patient.
Obtained using the \CRANpkg{OpenML} package. Contains 538 observations and 10
features. Target column is \code{"diseased"}.
}
\section{Pre-processing}{

\itemize{
\item All variable have been renamed.
\item The target variable has been re-encoded to \verb{"yes" and }"no"`.
}
}

\examples{
data("ilpd", package = "mlr3data")
str(ilpd)
}
\keyword{data}
