% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3data-package}
\alias{mlr3data}
\alias{mlr3data-package}
\title{mlr3data: Collection of Machine Learning Data Sets for 'mlr3'}
\description{
A small collection of interesting and educational machine learning data sets which are used as examples in the 'mlr3' book (\url{https://mlr3book.mlr-org.com}), the use case gallery (\url{https://mlr3gallery.mlr-org.com}), or in other examples. All data sets are properly preprocessed and ready to be analyzed by most machine learning algorithms. Data sets are automatically added to the dictionary of tasks if 'mlr3' is loaded.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3data}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3data/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})
}

Other contributors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID}) [contributor]
}

}
