% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moneyball.R
\docType{data}
\name{moneyball}
\alias{moneyball}
\alias{mlr_tasks_moneyball}
\title{Major League Baseball Statistics 1962-2012}
\source{
\url{https://www.openml.org/d/41021}
}
\description{
Regression data to predict the number of runs scored.
Obtained using the \CRANpkg{mlr3oml} package.

Contains 14 features and 1232 observations.
Target column is \code{"rs"}.
}
\section{Pre-processing}{

\itemize{
\item All variable names have been converted from upper case to lower case.
\item The variables \code{"year"}, \verb{"rs", }"ra"\verb{, }"w"` have been coerced to integers.
}
}

\examples{
data("moneyball", package = "mlr3data")
str(moneyball)
}
\keyword{data}
