% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optdigits.R
\docType{data}
\name{optdigits}
\alias{optdigits}
\alias{mlr_tasks_optdigits}
\title{Optical Recognition of Handwritten Digits}
\source{
\url{https://www.openml.org/d/980}
}
\description{
Classification data to predict handwritten digits.
Obtained using the \CRANpkg{mlr3oml} package.

Binarized version of the original data set. The multi-class target column has been converted to
a two-class nominal target column by re-labeling the majority class as positive (\code{"P"}) and all
others as negative (\code{"N"}). Originally converted by Quan Sun.

Contains 64 features and 5620 observations.
Target column is \code{"binaryclass"}.
}
\section{Pre-processing}{

\itemize{
\item All feature variables \code{"input1"}, ..., \code{"input64"} (number of on pixels in each block) have
been coerced to integers.
\item The target variable has been renamed from \code{"binaryClass"} to \code{"binaryclass"}.
}
}

\examples{
data("optdigits", package = "mlr3data")
str(optdigits)
}
\keyword{data}
