% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\alias{mlr_tasks_penguins}
\title{Palmer Penguins}
\source{
\CRANpkg{palmerpenguins}
}
\description{
Classification data to predict the species of penguins from the \CRANpkg{palmerpenguins} package
(CRAN version 0.1.0).
Promoted as an alternative to the \link[=iris]{iris data set}.
}
\section{Pre-processing}{

\itemize{
\item The unit have been removed from the column names.
Lengths are given in millimeters (mm), weight in gram (g).
}
}

\examples{
data("penguins", package = "mlr3data")
str(penguins)
}
\references{
\cite{gorman2014}

\url{https://github.com/allisonhorst/palmerpenguins}
}
\keyword{data}
