% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.R
\docType{data}
\name{titanic}
\alias{titanic}
\alias{mlr_tasks_titanic}
\title{Titanic}
\source{
\CRANpkg{titanic} and \url{https://www.kaggle.com/c/titanic/data}
}
\description{
Classification data to predict the fate of passengers on the ocean liner "Titanic".
Contains 10 features and 1309 observations. Target column is \code{"Survived"}.
}
\section{Pre-processing}{

\itemize{
\item All column names have been changed to \code{snake_case}.
\item training and test set have been joined.
Observations of the test set have a missing value in the target column \code{"survived"}.
\item Column '"survived"' has been re-encoded to a factor with levels '"yes"' and '"no"'.
\item Id column has been removed.
\item Passenger class \code{"pclass"} has been converted to an ordered factor.
\item Features \code{"sex"} and \code{"embarked"} have been converted to factors.
\item Empty strings in \code{"cabin"} and \code{"embarked"} have been encoded as missing values.
}
}

\examples{
data("titanic", package = "mlr3data")
str(titanic)
}
\keyword{data}
